==> 02_scripts/01_detailed_data/01b_harmonization.Rmd <==
---
title: "harmonization"
output: html_document
editor_options: 
  chunk_output_type: console
---

```{r , include = FALSE}

## Harmonization includes:
  ## Removing empty rows
  ## Check for various variables across all sheets if harmonized values are available for respective entries (e.g. country_id, material_id, unit_id)
  ## Actual harmonization of necessary values (e.g. country_id, material_id, unit_id)
  ## Additional adjustment (e.g. automatic filling of specific empty cells)
  ## Check if materials are inserted in "right" tables

## Harmonization is dependent on essential information provided in file sheets_columns.csv.


```

```{r setup , include = FALSE, echo = FALSE, message = FALSE}
knitr::opts_chunk$set(echo = TRUE)

library(tidyverse)
library(readxl)
library(knitr)
library(rlang)
library(DT)

# clear R environment
rm(list = ls())

```

<br> 

```{r , include = FALSE}
#### read input tables

# list of supposed column order
sheets_columns <- read_delim("./01_input/02_lists_and_concordance_tables/01_detailed_data/sheets_columns.csv", delim = ";")

# all other input tables
mat_ids <- read_delim("./01_input/02_lists_and_concordance_tables/material_ids.csv", delim = ";")

source_mat_ids <- read_delim("./01_input/02_lists_and_concordance_tables/source_material_ids.csv", delim = ";")

source_unit_ids <- read_delim("./01_input/02_lists_and_concordance_tables/source_unit_ids.csv", delim = ";")

source_country_ids <- read_delim("./01_input/02_lists_and_concordance_tables/source_country_ids.csv", delim = ";")

country_ids <- read_delim("./01_input/02_lists_and_concordance_tables/country_ids.csv", delim = ";")

poss_ent <- read_delim("./01_input/02_lists_and_concordance_tables/01_detailed_data/possible_entries.csv", delim = ";")

source_ids <- read_delim("./01_input/02_lists_and_concordance_tables/source_ids.csv", delim = ";")

log_diff_sources <- read_delim("./01_input/03_log_files/01_detailed_data/log_different_sources.csv", delim = ";")


# path to Excel data file
path <- "./01_input/01_data/01_detailed_data/detailed_data_mining.xlsx"

```

<br> 

```{r , echo = FALSE, warning = FALSE}

#### read excel with assigned column types 

# create list of supposed column types
type_list <- split(sheets_columns$supposed_type, sheets_columns$sheet)

# info on warnings()
  # NA entries are kept in Excel to keep reported NAs there for checking
  # Warnings occur for columns processing.mine_site_processing_incl_purchased, however there is no reason for these warnings

# read all excel sheets at once in a list, provide sheets names as list items
  # and allocate column types
detailed <- path %>%
  excel_sheets() %>%
  setdiff(., c("non-fitting_info_commodities", "environmental_concerns")) %>%
  set_names() %>% 
  map(~ read_excel(path = path, sheet = .x, col_types = type_list[[.x]], na = c("", "-", "NA")))

print(paste("Number of warnings during loading of excel file: ", length(warnings())))

```

<br> 

### Remove empty rows

```{r , echo = FALSE}

# for each sheet create an index which rows are all NA and remove them
  # as it creates problems with further processing otherwise

for (i in names(detailed)) {
  
  sheet <- detailed[[i]]
  
  ind <- apply(sheet, 1, function(x) all(is.na(x)))
  
  sheet <- sheet[ !ind, ]
  
  detailed[[i]] <- sheet

}

```

<br> 

### Overview of number of rows per table

```{r , echo = FALSE}

nr_rows <- map(detailed, nrow) %>% 
  as_tibble() %>% 
  pivot_longer(names(.)) %>% 
  rename(sheet = name, n = value)

# table for Rmd output
datatable(nr_rows, rownames = FALSE, caption = "Number of data rows per sheet")

```

<br> 

### Check of missing country_id allocations & of double entries for source_country_ids

```{r , echo = FALSE}

# create no_match table for countries
no_match_country <- tibble(
  country = setdiff(
    detailed$general$country %>%
      tibble() %>% 
      rename("country" = "."), 
    source_country_ids %>%
      filter(source_cat == "detailed") %>%
      select(source_country) %>% 
      rename("country" = "source_country")
    ) %>%
    pull(country)) %>% 
  filter(!is.na(country)) %>% 
  distinct()


# save no_match table as csv
if(nrow(no_match_country)>0){
write_delim(no_match_country, "./04_output/01_detailed_data/01_harmonization/01_no_match_tables/no_match_country.csv", delim = ";")
}

# table for Rmd output
datatable(no_match_country, rownames = FALSE, caption = "Missing source_country entries")



# check for double entries in source_mat_ids
double_source_cou <- source_country_ids %>%
  filter(source_cat == "detailed") %>%
  group_by(source_country) %>%
  summarise(n = n()) %>%
  filter(n > 1)

# table for Rmd output
datatable(double_source_cou, rownames = FALSE, caption = "Double entries in source_cou_ids (Please delete)")

```

<br> 

### Check of missing material_id allocations across all sheets & of double entries for source_mat_ids

```{r , echo = FALSE}

if(exists("source_materials")) rm(source_materials)

# dummy table for loop
source_materials <- tibble(sheet = NA, variable = NA, material = NA) %>% 
  filter(!is.na(material)) %>%
  mutate_all(as.character)

# list of all variables relevant for material_ids
sel_col <- sheets_columns %>% filter(check == "material_id")

# create table with entries of all material_ids across sheets
for (i in 1:nrow(sel_col)) {
  
  sheet_name <- sel_col[i,1] %>% pull()
  variable_name <- sel_col[i,2] %>% pull()

  a <- tibble(
    sheet = sheet_name,
    variable = variable_name,
    material = detailed[[sheet_name]] %>%
      select(all_of(variable_name)) %>%
      distinct() %>%
      filter(!is.na(.)) %>%
      pull()
    )

  source_materials <- source_materials %>%
    dplyr::union(., a)
  
  }

# material_ids entries not yet in material_id table (or not relevant for allocation)
no_match_mat <- source_materials %>% 
  filter(!(material %in% (source_mat_ids %>% 
                            filter(source_cat == "detailed") %>% 
                            pull(source_material_id)
                          )
           )
         )

#create directory
dir.create("./04_output/01_detailed_data/01_harmonization/01_no_match_tables/", showWarnings = FALSE)

# save no_match table as csv
write_delim(no_match_mat, "./04_output/01_detailed_data/01_harmonization/01_no_match_tables/no_match_mat.csv", delim = ";")

# table for Rmd output
datatable(no_match_mat, rownames = FALSE, caption = "Missing source_material entries")



# check for double entries in source_mat_ids
double_source_mat <- source_mat_ids %>%
  filter(source_cat == "detailed") %>%
  group_by(source_material_id) %>%
  summarise(n = n()) %>%
  filter(n > 1)

# table for Rmd output
datatable(double_source_mat, rownames = FALSE, caption = "Double entries in source_mat_ids (Please delete)")

```

<br> 

### Check of missing unit_id allocations across all sheets & of double entries for source_unit_ids

```{r , echo = FALSE}

if(exists("source_units")) rm(source_units)

# dummy table for loop
source_units <- tibble(sheet = NA, variable = NA, unit = NA) %>% 
  filter(!is.na(unit)) %>%
  mutate_all(as.character)

# list of all variables relevant for unit_ids
sel_col <- sheets_columns %>% filter(check %in% c("unit_id", "grade_unit_id"))

# create table with entries of all unit_ids across sheets
for (i in 1:nrow(sel_col)) {
  
  sheet_name <- sel_col[i,1] %>% pull()
  variable_name <- sel_col[i,2] %>% pull()

  a <- tibble(
    sheet = sheet_name,
    variable = variable_name,
    unit = detailed[[sheet_name]] %>%
      select(all_of(variable_name)) %>%
      distinct() %>%
      filter(!is.na(.)) %>%
      pull()
  )

  source_units <- source_units %>%
    dplyr::union(., a)
  
}

# source_unit_id entries not yet in source_unit_id table (or not relevant for allocation)
no_match_unit <- source_units %>%
  filter(!(unit %in% (source_unit_ids %>% 
                            filter(source_cat == "detailed") %>% 
                            pull(source_unit)
                          )
           )
         )


# save no_match table as csv
write_delim(no_match_unit, "./04_output/01_detailed_data/01_harmonization/01_no_match_tables/no_match_unit.csv", delim = ";")

# table for Rmd output
datatable(no_match_unit, rownames = FALSE, caption = "Missing source_unit_id entries")



# check for double entries in source_unit_ids
double_source_unit <- source_unit_ids %>%
  filter(source_cat == "detailed") %>%
  group_by(source_unit) %>%
  summarise(n = n()) %>%
  filter(n > 1)

# table for Rmd output
datatable(double_source_unit, rownames = FALSE, caption = "Double entries in source_unit_ids (Please delete)")

```

<br> 

### Check of possible entries across all sheets & of double entries in pre-defined table for possible entries

```{r , echo = FALSE}

if(exists("source_ent")) rm(source_ent)

# dummy table for loop
source_ent <- tibble(variable = NA, entry = NA) %>% 
  filter(!is.na(entry)) %>%
  mutate_all(as.character)

# list of all variables relevant for possible entries
sel_col <- sheets_columns %>% filter(check == "poss_ent")

# create table with entries of all entries in respective columns across sheets
for (i in 1:nrow(sel_col)) {
  
  sheet_name <- sel_col[i,1] %>% pull()
  variable_name <- sel_col[i,2] %>% pull()

  a <- tibble(
    variable = variable_name,
    entry = detailed[[sheet_name]] %>%
      select(all_of(variable_name)) %>%
      distinct() %>%
      filter(!is.na(.)) %>%
      pull()
  )

  source_ent <- source_ent %>%
    dplyr::union(., a)
  
}

# entries not yet included in table for possible entries (or not relevant)
no_match_entries <- source_ent %>% filter(!(entry %in% poss_ent$source_entry))

# save no_match table as csv
write_delim(no_match_entries, "./04_output/01_detailed_data/01_harmonization/01_no_match_tables/no_match_entries.csv", delim = ";")

# table for Rmd output
datatable(no_match_entries, rownames = FALSE, caption = "Check entries and add to table of possible entries")



# check for double entries in table for possible entries
double_poss_ent <- poss_ent %>%
  group_by(source_entry) %>%
  summarise(n = n()) %>%
  filter(n > 1)

# table for Rmd output
datatable(double_poss_ent, rownames = FALSE, caption = "Double entries in possible entries (Please delete)")

```

<br> 

### Check of number of entries for each mine_fac in sheet “general”

```{r , echo = FALSE}

nr_mines <- detailed$general %>%
  group_by(mine_fac) %>%
  summarise(n = n()) %>%
  filter(n > 1)

# table for Rmd output
datatable(nr_mines, rownames = FALSE, caption = "Mines with more than one entry")

```

<br> 

### Check of entries for mine_fac names across all sheets

```{r , echo = FALSE}

if(exists("mine_names")) rm(mine_names)

# dummy table for loop
mine_names <- tibble(sheet = NA, variable = NA, mine_fac = NA) %>%
  filter(!is.na(mine_fac)) %>%
  mutate_all(as.character)

# list of all variables relevant for mine_fac
sel_col <- sheets_columns %>% filter(check == "mines")

# create table with entries of all mine_fac across sheets
for (i in 1:nrow(sel_col)) {

  sheet_name <- sel_col[i,1] %>% pull()
  variable_name <- sel_col[i,2] %>% pull()

  a <- tibble(
    sheet = sheet_name,
    variable = variable_name,
    mine_fac = detailed[[sheet_name]] %>%
      select(all_of(variable_name)) %>%
      distinct() %>%
      filter(!is.na(.)) %>%
      pull()
  )

  mine_names <- mine_names %>%
    dplyr::union(., a)

}

# mine_fac names which are different from sheet 'general'
no_match_mines <- mine_names %>% filter(!(mine_fac %in% detailed$general$mine_fac))

# save no_match table as csv
write_delim(no_match_mines, "./04_output/01_detailed_data/01_harmonization/01_no_match_tables/no_match_mines.csv", delim = ";")

# table for Rmd output
datatable(no_match_mines, rownames = FALSE, caption = "Incorrect mine_fac names (different than on sheet 'general')")

```

<br> 

### Check of entries for sub-site names across all sheets

```{r , echo = FALSE}

if(exists("sub_sites")) rm(sub_sites)

# dummy table for loop
sub_sites <- tibble(sheet = NA, variable = NA, sub_site = NA) %>%
  filter(!is.na(sub_site)) %>%
  mutate_all(as.character)

# list of all variables relevant for sub_sites
sel_col <- sheets_columns %>% filter(check == "sub_sites")

# create table with entries of all sub_sites across sheets
for (i in 1:nrow(sel_col)) {

  sheet_name <- sel_col[i,1] %>% pull()
  variable_name <- sel_col[i,2] %>% pull()

  a <- tibble(
    sheet = sheet_name,
    variable = variable_name,
    sub_site = detailed[[sheet_name]] %>%
      select(all_of(variable_name)) %>%
      distinct() %>%
      filter(!is.na(.)) %>%
      pull()
  )

  sub_sites <- sub_sites %>%
    dplyr::union(., a)

}

# sub-site names which are different from sheet 'sub_sites'
no_match_sub_sites <- sub_sites %>% filter(!(sub_site %in% detailed$sub_sites$`sub_site`))

# save no_match table as csv
write_delim(no_match_sub_sites, "./04_output/01_detailed_data/01_harmonization/01_no_match_tables/no_match_sub_sites.csv", delim = ";")

# table for Rmd output
datatable(no_match_sub_sites, rownames = FALSE, caption = "Incorrect sub-site names (different than on sheet 'sub_sites)'")

```

<br> 

### Check of entries for years across all sheets

```{r , echo = FALSE}

if(exists("year_ranges")) rm(year_ranges)

# dummy table for loop
years <- tibble(sheet = NA, variable = NA, year = NA) %>%
  filter(!is.na(year)) %>%
  mutate_at(vars(sheet, variable), as.character) %>%
  mutate_at(vars(year), as.double)

# list of all variables relevant for years
sel_col <- sheets_columns %>% filter(check == "range")

# create table with entries of all years across sheets
for (i in 1:nrow(sel_col)) {

  sheet_name <- sel_col[i,1] %>% pull()
  variable_name <- sel_col[i,2] %>% pull()

  a <- tibble(
    sheet = sheet_name,
    variable = variable_name,
    year = detailed[[sheet_name]] %>%
      select(all_of(variable_name)) %>%
      distinct() %>%
      filter(!is.na(.)) %>%
      pull()
  )

  years <- years %>%
    dplyr::union(., a)

}

# years which are lower than 1970 or higher than the current year (i.e. years which don't make much sense)
no_match_years <- years %>% filter(year < 1970 | year > as.double(substr(Sys.Date(), 1, 4)))

# save no_match table as csv
write_delim(no_match_years, "./04_output/01_detailed_data/01_harmonization/01_no_match_tables/no_match_years.csv", delim = ";")

# table for Rmd output
datatable(no_match_years, rownames = FALSE, caption = "Incorrect years")

```

<br> 

### Check of NAs where no NA possible

```{r , echo = FALSE}

if(exists("source_na")) rm(source_na)

# dummy table for loop
source_na <- tibble(sheet = NA, variable = NA) %>% 
  filter(!is.na(variable)) %>%
  mutate_all(as.character)

# list of all variables relevant for material_ids
sel_col <- sheets_columns %>% filter(na_possible == "no")

# create table with entries of all material_ids across sheets
for (i in 1:nrow(sel_col)) {
  
  sheet_name <- sel_col[i,1] %>% pull()
  variable_name <- sel_col[i,2] %>% pull()
  
  a <- detailed[[sheet_name]] %>%
    filter(is.na(eval(parse(text = variable_name)))) %>%
    mutate(sheet = sheet_name, variable = variable_name) %>%
    distinct(sheet, variable)

  source_na <- source_na %>%
    dplyr::union(., a)
  
}

# save no_match table as csv
write_delim(source_na, "./04_output/01_detailed_data/01_harmonization/01_no_match_tables/source_na.csv", delim = ";")

# table for Rmd output
datatable(source_na, rownames = FALSE, caption = "NAs where no NA possible")

```

<br> 

### Check for same mine having entries by more than one company in a year
#### (excluding those from respective log file)

```{r , echo = FALSE, message = FALSE}

if(exists("double_source")) rm(double_source)

# dummy table for loop
double_source <- tibble(sheet = NA, mine_fac = NA, year = NA, n = NA) %>% 
  filter(!is.na(sheet)) %>%
  mutate(
    across(c(sheet, mine_fac), as.character),
    across(c(year, n), as.integer)
    )

# list of all variables relevant for sources
sel_col <- sheets_columns %>% filter(check %in% c("source"))
i <- 1
# create table with all mine_fac with double source per year across sheets
for (i in 1:nrow(sel_col)) {
  
  sheet_name <- sel_col[i,1] %>% pull()
  
  a <- detailed[[sheet_name]] %>%
    mutate(source_check = substring(source, 1, 5)) %>%
    distinct(mine_fac, year, source_check) %>%
    group_by(mine_fac, year) %>%
    summarise(n = n()) %>%
    ungroup() %>%
    filter(n > 1) %>%
    add_column(sheet = sheet_name, .before = "mine_fac") %>%
    filter(!(mine_fac %in% (log_diff_sources %>% filter(sheet == sheet_name) %>% pull(mine_fac))))

  double_source <- double_source %>%
    dplyr::union(., a)
  
}


# table for Rmd output
datatable(double_source, rownames = FALSE, caption = "Unchecked mine_fac with entry from more than one company per year")

```

<br> 

### Harm. of country_ids (only sheet general)

```{r , echo = FALSE}

general <- detailed$general

# left join table with alphanumiso, select alphanumiso and put it at the proper position
general <- general %>%
  add_column(
    general %>% 
      left_join(., 
                source_country_ids %>%
                  filter(source_cat == "detailed") %>%
                  select(source_country, alphanumiso), 
                by = c("country" = "source_country")
                ) %>% 
      select(alphanumiso),
    .after = "country"
  )

# harmonize different country names by inserting country names on alphanumiso 
general <- general %>% 
  select(-country) %>%
  left_join(country_ids %>% 
              select(alphanumiso, country),
            by = "alphanumiso") %>% 
  relocate(country, .before = alphanumiso)

detailed$general <- general

print("Done!")

```

<br> 

### Harm. of material_ids across all sheets

```{r , echo = FALSE}

sel_col <- sheets_columns %>% filter(check == "material_id")

# in a loop take each sheet which needs material_ids, do a left join with material_ids, select material_ids and put them at the proper position
for (i in 1:nrow(sel_col)) {
  
  sheet_name <- sel_col[i,1] %>% pull()
  variable_name <- sel_col[i,2] %>% pull()
  
  sheet <- detailed[[sheet_name]]
  
  # replaces the respective material column with id
    # (in contrary to just adding id, which has not been chosen because of various reasons)
    # (however, if names should be kept due to some reason, code could be adjusted similar to adding alphanumiso above)
  sheet <- sheet %>%
    left_join(.,
              source_mat_ids %>%
                filter(source_cat == "detailed") %>%
                select(source_material_id, material_id),
              by=setNames(nm=variable_name, "source_material_id")
              ) %>%
    mutate(!!variable_name := material_id) %>%
    select(-material_id)
  
  detailed[[sheet_name]] <- sheet

}

print("Done!")

```

<br> 

### Harm. of unit_ids across all sheets

```{r , echo = FALSE}

sel_col <- sheets_columns %>% filter(check %in% c("unit_id", "grade_unit_id"))

# in a loop take each sheet which needs unit_ids, do a left join with unit_ids, select unit_ids and put them at the proper position
for (i in 1:nrow(sel_col)) {
  
  sheet_name <- sel_col[i,1] %>% pull()
  variable_name <- sel_col[i,2] %>% pull()
  
  sheet <- detailed[[sheet_name]]
  
  # replace unit entries with respective IDs
  sheet <- sheet %>%
    left_join(.,
              source_unit_ids %>%
                filter(source_cat == "detailed") %>%
                select(source_unit, unit_id), by=setNames(nm=variable_name, "source_unit")
              ) %>%
    mutate(!!variable_name := unit_id) %>%
    select(-unit_id)
  
  detailed[[sheet_name]] <- sheet
  
}

print("Done!")

```

<br> 

### Harm. of possible entries across all sheets

```{r , echo = FALSE}

sel_col <- sheets_columns %>% filter(check == "poss_ent")

# in a loop take each sheet which needs replacement, do a left join with harmonized entry_names, select entry_names and put them at the proper position by replacing respective column
for (i in 1:nrow(sel_col)) {
  
  sheet_name <- sel_col[i,1] %>% pull()
  variable_name <- sel_col[i,2] %>% pull()
  
  sheet <- detailed[[sheet_name]]
  
  # replace unit entries with respective IDs
  sheet <- sheet %>%
    left_join(.,
              poss_ent %>%
                select(source_entry, entry_name), by=setNames(nm=variable_name, "source_entry")
              ) %>%
    mutate(!!variable_name := entry_name) %>%
    select(-entry_name)
  
  detailed[[sheet_name]] <- sheet
  
}

print("Done!")

```

<br> 

### Harm. of source_ids across all sheets

```{r , echo = FALSE, message = FALSE}

# replace entries for sources

# check and create ids
# make distinct table of all, deselect all which already have an id in id table, then add id for each one left

for (i in names(detailed)) {
  
  sheet <- detailed[[i]]
  
  # sources of current sheet
  sources <- sheet %>% distinct(source, source_url)
  
  # get the highest number from current source_ids
  last_nr <- source_ids$source_id %>% substr(., 5, 8) %>% as.integer() %>% max()
  
  # get new sources
  new_sources <- setdiff(sources, source_ids %>% select(source, source_url))
  
  # provide new sources with ids
  if(nrow(new_sources) > 0) {
    new_sources <- new_sources %>%
      mutate(source_id = paste0(
        "det_",
        seq(last_nr + 1, last_nr + nrow(new_sources), by=1)
        )
        )
  
  source_ids <- source_ids %>%
    union(., new_sources)
  
  }
  
  # replace sources with respective IDs
  sheet <- sheet %>%
    left_join(., source_ids) %>%
    select(-source, -source_url)
  
  detailed[[i]] <- sheet
  
}

print("Done!")
  

# save new source_ids
write_delim(source_ids, "./01_input/02_lists_and_concordance_tables/source_ids.csv",
            delim = ";", na = "")

```

<br>

### Check if material_ids are included in right table 

```{r , echo = FALSE}

# Based on material_categories in material_ids.csv. E.g. column "commodity" is supposed to only contain those material_ids associated with material_cat "commodities"

## For minerals

# Vector with IDs
mineral_sheet_ids <- c(
  mat_ids %>%
    filter(material_category %in% c("mineral", "ff_mineral") | 
           material_category_2 %in% c("ore concentrate", "metal ore product")) %>%
    distinct(material_id) %>%
    pull(),
  mat_ids %>%
    filter(material_category %in% c("mineral", "ff_mineral") | 
           material_category_2 %in% c("ore concentrate", "metal ore product")) %>%
    distinct(material_id_agg) %>%
    pull(material_id_agg)
  )

mineral_sheet_ids <- mineral_sheet_ids[!is.na(mineral_sheet_ids)]

sheet <- detailed$minerals_ores_conce

wrong_min_entries <- sheet %>%
  filter(!(min_ore_con %in% mineral_sheet_ids) & !is.na(min_ore_con)) %>%
  distinct(min_ore_con)

# table for Rmd output
datatable(wrong_min_entries, rownames = FALSE, caption = "Material IDs which should be in sheet 'commodity'")

## For commodities
comm_sheet_ids <- c(
  mat_ids %>%
    filter(material_category %in% c("commodity") | 
           material_category_2 %in% c("metal alloy", "metal product")) %>%
    distinct(material_id) %>%
    pull(),
  mat_ids %>%
    filter(material_category %in% c("commodity") | 
           material_category_2 %in% c("metal alloy", "metal product")) %>%
    distinct(material_id_agg) %>%
    pull(material_id_agg)
  )

comm_sheet_ids <- comm_sheet_ids[!is.na(comm_sheet_ids)]

sheet <- detailed$commodities

wrong_comm_entries <- sheet %>%
  filter(!(commodity %in% comm_sheet_ids) & !is.na(commodity)) %>%
  distinct(commodity)

# table for Rmd output
datatable(wrong_comm_entries, rownames = FALSE, caption = "Material IDs which should be in sheet 'minerals_ores_conce'")




## For table reserves
sheet <- detailed$capacity_reserves

# Material IDs in column minerals_and_ores, which should be in column commodity
wrong_res_entries_at_mat <- sheet %>%
  filter(!(minerals_and_ores %in% mineral_sheet_ids) & !is.na(minerals_and_ores)) %>% 
  select(mine_fac, sub_site, reserves_year, minerals_and_ores, commodity)

# table for Rmd output
datatable(wrong_res_entries_at_mat, rownames = FALSE, caption = "Material IDs in sheet capacity_reserves$minerals_and_ores which should be in capacity_reserves$commodity")


# Material IDs in column commodity, which should be in column minerals_and_ores
wrong_res_entries_at_comm <- sheet %>%
  filter(!(commodity %in% comm_sheet_ids) & !is.na(commodity)) %>% 
  select(mine_fac, sub_site, reserves_year, minerals_and_ores, commodity)

# table for Rmd output
datatable(wrong_res_entries_at_comm, rownames = FALSE, caption = "Material IDs in sheet capacity_reserves$commodity which should be in capacity_reserves$minerals_and_ores")




# Entry in reserves_mineral_value but without specifying the mineral
reserves_missing_mat <- sheet %>%
  filter(!is.na(reserves_mineral_value) & is.na(minerals_and_ores)) %>% 
  select(mine_fac, sub_site, reserves_year, minerals_and_ores, commodity, reserves_mineral_value, reserves_commodity_value)

# table for Rmd output
datatable(reserves_missing_mat, rownames = FALSE, caption = "Entry in reserves_mineral_value but the mineral is not specified")



# Entry in reserves_commodity_value but without specifying the commodity
reserves_missing_comm <- sheet %>%
  filter(!is.na(reserves_commodity_value) & is.na(commodity)) %>% 
  select(mine_fac, sub_site, reserves_year, minerals_and_ores, commodity, reserves_mineral_value, reserves_commodity_value)

# table for Rmd output
datatable(reserves_missing_comm, rownames = FALSE, caption = "Entry in reserves_commodity_value but the commodity is not specified")


# Entry in reserves_commodity_value but without specifying the commodity
reserves_missing_comm_grade <- sheet %>%
  filter(!is.na(grade) & is.na(commodity)) %>% 
  select(mine_fac, sub_site, reserves_year, minerals_and_ores, commodity, reserves_mineral_value, reserves_commodity_value)

# table for Rmd output
datatable(reserves_missing_comm_grade, rownames = FALSE, caption = "Entry in grade but the  commodity is not specified")

```

<br> 

```{r , echo = FALSE, message = FALSE}

#### Additional adjustments for harmonization

sheet_min <- detailed$minerals_ores_conce

sheet_com <- detailed$commodities


## Adjustment of "O.bulk" and "Con.bulk"

# all combinations (mine_fac, sub_site, type_mining, year) with more than one commodity/metal where min_ore_con is NA
several_com <- sheet_com %>%
  filter(is.na(min_ore_con)) %>%
  group_by(mine_fac, sub_site, type_mining, year) %>%
  summarise(n = n()) %>%
  filter(n > 1) %>%
  select(-n) %>%
  ungroup() %>%
  mutate(bulk = "yes")

sheet_min <- sheet_min %>%
  left_join(., several_com) %>%
  mutate(min_ore_con = ifelse(
    is.na(min_ore_con) &
            type_mineral %in% c("Ore mined", "Ore processed") &
            bulk == "yes",
    "O.bulk",
    min_ore_con
  )) %>%
  select(-bulk)

sheet_com <- sheet_com %>%
  left_join(., several_com) %>%
  mutate(min_ore_con = ifelse(
    is.na(min_ore_con) &
            bulk == "yes",
    "O.bulk",
    min_ore_con
  )) %>%
  select(-bulk)




## Adjust type_mineral -> make NA = "Ore processed" and make other NA = "NM mineral"
sheet_min <- sheet_min %>%
  mutate(type_mineral = ifelse(
    is.na(type_mineral) &
      min_ore_con %in% (
        na.omit(
        c(
          mat_ids[mat_ids$material_category_2 %in%
                    c("metal ore", "non-metallic mineral ore"),]$material_id,
          mat_ids[mat_ids$material_category_2 %in%
                    c("metal ore", "non-metallic mineral ore"),]$material_id_agg
          )
        )
      ),
    "Ore processed",
    type_mineral
  )
  ) %>% 
  mutate(type_mineral = ifelse(
    is.na(type_mineral) &
      min_ore_con %in% (
        na.omit(
        c(
          mat_ids[mat_ids$material_category %in%
                    c("mineral"),]$material_id,
          mat_ids[mat_ids$material_category %in%
                    c("mineral"),]$material_id_agg
          ) # (as ores have already been checked/filled, here only the first material_category is used, as only extraction should be included in this table)
        )
      ),
    "NM mineral",
    type_mineral
  )
  )



detailed$minerals_ores_conce <- sheet_min

detailed$commodities <- sheet_com

```

<br> 

```{r , echo = FALSE}

#### save harmonized tables

write_rds(detailed, "./03_intermediate/01_detailed_data/all_tables_harmonized.rds")

```

<br>
