==> 02_scripts/01_detailed_data/07_georeferencing.Rmd <==
---
title: "Georeferencing"
output: html_document
editor_options: 
  chunk_output_type: console
---


###### In this script, mines are georeferenced 
##### 1. Assignment of facility_ids
##### 2. Conversion of sheet_general to sf objects
##### 3. georeferencing of mines that have coordinates in sheet_general or sheet_sub_sites
##### 3.1 assign GADM codes (GIDs) to mines with coords by intersecting coordinates with GADM Polygons
##### 4. georeferencing of mines and regions that do not have coordinates in sheet_general or sheet_sub_sites
##### 4.1. check if there are missing regions in the WU_GADM_concordance.xlsx file, if so update manually
##### 4.2.1. add GADM regions to mines without coordinates
##### 4.2.2. OPTIONAL: add polygons to mines without coordinates
##### 5. Export final data



### Read In data
```{r , echo = FALSE, message = TRUE} 

# clear R environment
rm(list = ls())

library(tidyverse)
library(sf)
library(mapview)

# read harmonized data file (detailed data)
detailed <- read_rds("./03_intermediate/01_detailed_data/gaps_filled.rds")
general <- detailed$general
sub_sites <- detailed$sub_sites

# Import Geodata from GADM if necessary
source("./02_scripts/05_other/get_GADM_data.R")

file <- "./01_input/06_geodata/gadm36_levels.gpkg"
st_layers(file)
data <- src_sqlite(file)

gadm_combined <- read_csv("./01_input/06_geodata/GADM_levels_combined.csv", show_col_types = FALSE)

log_GADM <- read_csv("./01_input/03_log_files/01_detailed_data/log_location_not_intersecting_with_GADM.csv", show_col_types = FALSE)

if(file.exists("./01_input/02_lists_and_concordance_tables/01_detailed_data/WU_GADM_log_concordance.csv")) {
  WU_GADM_log_concordance <- read_csv("./01_input/02_lists_and_concordance_tables/01_detailed_data/WU_GADM_log_concordance.csv", show_col_types = FALSE)
} else {
  WU_GADM_log_concordance <- tibble(facility_id = chr(), 
                                    GID_0 = chr(), 
                                    GID_1 = chr(), 
                                    GID_2 = chr(), 
                                    GID_3 = chr(), 
                                    GID_4 = chr())
}


# load functions on which this script depends
source("./02_scripts/00_functions/split_columns.R")
source("./02_scripts/00_functions/plot_polygons.R")

# set this variable to TRUE if polygons should be assigned to mines without coordinates
# This assignment of polygons is time and storage intensive
# if it is set to FALSE, only the GADM levels of mines without coordinates are assigned, 
# which can then be linked to an external GADM table
# by default, it should be FALSE
add_polygons = FALSE

```


<br>

### 1. Assignment of facility_id
```{r , echo = FALSE, message = FALSE} 

# create sheet containing all mines and sub-sites
general <- bind_rows(general, sub_sites)


# the only column that has a different naming in the two sheets but refers to the same is:
# mining_facility_types (in sheet_general) and mine_types (in sheet_sub_sites).
# therefore, they are put together into one column, named mining_facility_types
general$mining_facility_types <- coalesce(general$mining_facility_types, general$mine_types)
general <- select(general, -mine_types)

# just for testing purposes
# general <- general %>%
#   arrange(mine_fac) %>%
#   slice(1001:2000)

# needed for check later
general_before <- general


if (file.exists("./01_input/02_lists_and_concordance_tables/facility_ids.csv")){
  
  # read facility_ids log table
  facility_ids <- read_csv("./01_input/02_lists_and_concordance_tables/facility_ids.csv")

  # 1. assign existing facility_ids to the corresponding facilities and sub-sites
  general_w_fac_id <- general %>%
    arrange(mine_fac) %>%
    left_join(facility_ids, by = c("mine_fac", "sub_site")) %>% 
    relocate(facility_id, .before = mine_fac) %>% 
    filter(!(facility_id %>% is.na()))
  
  
  # check if there are facilities without allocated facility_ids
  general_wo_fac_id <- general %>% 
    left_join(facility_ids, by = c("mine_fac", "sub_site")) %>% 
    filter(facility_id %>% is.na()) %>% 
    select(-facility_id)
  
  if (nrow(general_wo_fac_id) > 0 ) {
    
    # 2. assign new facility_ids to new entries, if necessary
    
    # 2.1. handle cases where new sub-sites of already existing parent facilities are introduced
    # they should have the first five digits of the parent facility, with a new unique two-digit ending
    facilities_with_parent_id <- facility_ids$mine_fac %>% unique()

    general_wo_fac_id_1 <- general_wo_fac_id %>% 
      filter(mine_fac %in% facilities_with_parent_id)
    
    if (nrow(general_wo_fac_id_1) > 0) {
        
      # assign parent's facility id
      general_wo_fac_id_1 <- general_wo_fac_id_1 %>% 
        left_join(facility_ids %>% 
                    mutate(facility_id = substr(facility_id, 1, 8)) %>%
                    select(-sub_site) %>% 
                    unique(), by = "mine_fac") %>% 
        relocate(facility_id, .before = mine_fac)
      
      
      # assign new ending
      general_wo_fac_id_1_temp <- tibble()
        
      for(row_nr in (1:nrow(general_wo_fac_id_1))) {
        
        row <- general_wo_fac_id_1[row_nr,]
        
        # get last sub-site number of this facility
        last_sub_nr <- facility_ids %>% 
          bind_rows(general_wo_fac_id_1_temp) %>% 
          filter(mine_fac == row$mine_fac) %>%
          select(facility_id) %>% 
          mutate(facility_id = substr(facility_id, 10,11)) %>%
          unlist() %>%
          as.vector() %>%
          as.integer() %>%
          max()
        
        # paste last two digits to facility id
        row <- row %>% 
          mutate(facility_id = str_c(facility_id, ".", str_pad(last_sub_nr + 1, width = 2, pad = "0", "left")))
        
        general_wo_fac_id_1_temp <- general_wo_fac_id_1_temp %>% 
          bind_rows(row)
        
      }
      
      # replace original object with version with added facility_ids with correct ending
      general_wo_fac_id_1 <- general_wo_fac_id_1_temp
      
    }
    
    
    # 2.2. handle cases where new facilities and sub-sites need new facility_ids
    general_wo_fac_id_2 <- general_wo_fac_id %>% 
      filter(!(mine_fac %in% facilities_with_parent_id))
    
    if(nrow(general_wo_fac_id_2) > 0) {
      
      last_nr <- facility_ids$facility_id %>% substr(., 4, 8) %>% as.integer() %>% max()
      
      general_wo_fac_id_2 <- general_wo_fac_id_2 %>%
        arrange(mine_fac) %>%
        group_by(mine_fac) %>% 
          mutate(facility_id = str_c("COM", str_pad(dplyr::cur_group_id() + last_nr, width = 5, pad = "0"))) %>%
        ungroup() %>%
        group_by(facility_id) %>% 
          mutate(facility_id = str_c(facility_id, ".", str_pad(row_number() - 1, width = 2, pad = "0", "left"))) %>%
        ungroup() %>%
        relocate(facility_id, .before = mine_fac)
      
    }
    
    # 3. combine all entries that got a new facility_id assigned
    if (nrow(general_wo_fac_id_1) > 0 & nrow(general_wo_fac_id_2) > 0){
      general_wo_fac_id <- general_wo_fac_id_1 %>% 
        bind_rows(general_wo_fac_id_2) %>% 
        arrange(facility_id)
      
    } else if (nrow(general_wo_fac_id_1) > 0) {
      general_wo_fac_id <- general_wo_fac_id_1 %>%
        arrange(facility_id)
      
    } else if (nrow(general_wo_fac_id_2) > 0) {
      general_wo_fac_id <- general_wo_fac_id_2 %>%
        arrange(facility_id)
    }
    
    
    # 4. add the new facility_ids to the facility_ids.csv file 
    facility_ids <- facility_ids %>% 
      bind_rows(general_wo_fac_id %>%
                  select(facility_id, mine_fac, sub_site))
    
    write_csv(facility_ids, "./01_input/02_lists_and_concordance_tables/facility_ids.csv", na = "")
    
    
    # 5. create table general which now has all facility_ids assigned
    general <- bind_rows(general_w_fac_id, general_wo_fac_id) %>%
      arrange(facility_id)
    print("New facility_ids were added to the existing file facility_ids.csv")

  
    # 6. check that general has the same number of rows as before the assignment of facility_ids
    if(nrow(general) != nrow(general_before)){
      
      print(setdiff(general %>%
                      select(mine_fac, sub_site),
                    general_before %>%
                      select(mine_fac, sub_site)))
      
      print(setdiff(general_before %>%
                      select(mine_fac, sub_site),
                    general %>%
                      select(mine_fac, sub_site)))
      
      # duplicated names
      print(general_before %>% 
              filter(duplicated(general_before %>% select(mine_fac, sub_site)) == TRUE))
      
      stop(cat("The assignment of facility_ids caused the table general (nrow: ", nrow(general), 
               ") to have more or less rows than it should have (nrow:", nrow(general_before), ")"))
    }
    
    
  } else {
    
    general <- general_w_fac_id
    print("All facility_ids were allocated from the file facility_ids.csv, no new facility_ids were added")
  }
  
} else {
  
  # create new facility_id for all facilities
  general <- general %>% 
    arrange(mine_fac) %>% 
    group_by(mine_fac) %>% 
      mutate(facility_id = str_c("COM", str_pad(dplyr::cur_group_id(), width = 5, pad = "0"))) %>%
    ungroup() %>% 
    group_by(facility_id) %>% 
      mutate(facility_id = str_c(facility_id, ".", str_pad(row_number()-1, width = 2, pad = "0", "left"))) %>% 
    ungroup() %>%
    relocate(facility_id, .before = mine_fac)

  
  # Create new file facility_ids.csv with all facility_ids from mine-level data
  facility_ids <- general %>% 
    select(facility_id, mine_fac, sub_site)
  
  write_csv(facility_ids, "./01_input/02_lists_and_concordance_tables/facility_ids.csv", na = "")
  print("New facility_ids.csv file created")

}


```

<br>

### 2. Conversion of sheet_general to sf objects
```{r , echo = FALSE, message = FALSE} 

# If there are NA values in the latitude or longitute, add a column to keep information
general$coord_is_na <- ifelse(is.na(general$longitude), TRUE, FALSE)

general_w_coords <- general %>% 
  filter(., coord_is_na == FALSE)

general_wo_coords <- general %>% 
  filter(., coord_is_na == TRUE)


# convert their coordinates into sf objects
general_w_coords.sf <- st_as_sf(x = general_w_coords, 
                               coords = c("longitude", "latitude"),
                               crs = "+proj=longlat +datum=WGS84 +ellps=WGS84 +towgs84=0,0,0")

general_wo_coords.sf <- st_as_sf(x = select(general_wo_coords, -latitude, -longitude), 
                                 geometry = st_as_sfc("POINT EMPTY"), 
                                 crs = "+proj=longlat +datum=WGS84 +ellps=WGS84 +towgs84=0,0,0")

general.sf <- bind_rows(general_w_coords.sf, general_wo_coords.sf)


# divide  mines and processing into mines_general and mines_sub_sites again
# the reason is that mines_general can be multipoints, while mines_sub_sites cannot
mines_general <- filter(general.sf, is.na(sub_site))
mines_sub_sites <- filter(general.sf, !is.na(sub_site))

```


<br>

### 3. georeferencing of mines that have coordinates in sheet_general or sheet_sub_sites
```{r , echo = FALSE, message = FALSE, warning = FALSE} 

# mines that do have sub_sites with coordinates
mines_general_w_multipoints_list <- filter(mines_sub_sites, coord_is_na == FALSE)$mine_fac %>% unique()

mines_general_w_multipoints <- mines_general %>%
  filter(mine_fac %in% mines_general_w_multipoints_list)


# mines that do not have sub_sites with coordinates, but have coordinates in sheet general
mines_general_wo_multipoints <- mines_general %>%
  filter(!(mine_fac %in% mines_general_w_multipoints_list) & coord_is_na == FALSE)


# mines that do not have sub_sites with coordinate and do not have coordinates in sheet general
mines_general_wo_coords <- mines_general %>%
  filter(!(mine_fac %in% mines_general_w_multipoints_list) & coord_is_na == TRUE)


# this has to return TRUE
nrow(mines_general_wo_multipoints) + nrow(mines_general_w_multipoints) +
  nrow(mines_general_wo_coords) == nrow(mines_general)


# mines that do not have sub_sites with coordinates can just be added to general_output
general_output <- mines_general_wo_multipoints

# for those mines_general that have sub_sites with coordinates (mines_general_w_multipoints), create multipoints with all sub_sites
for(mine in mines_general_w_multipoints_list){
  
  #select a particular mine with all its sub_sites that have coordinates
  df <- filter(general.sf, mine_fac == mine & coord_is_na == FALSE)
  
  #combine all the points of a particular mine and its sub_sites to a multipoint
  multipoint <- df$geometry %>%
    st_difference() %>% #removes potential duplicate points
    st_combine() #combines the points to multipoint
  
  #replace the original point of the (general) mine with the multipoint
  mines_general_w_multipoints[mines_general_w_multipoints$mine_fac == mine & is.na(mines_general_w_multipoints$sub_site),]$geometry <- multipoint
  
  # set the column coord_is_na to FALSE, as now this mine has coordinates
  mines_general_w_multipoints[mines_general_w_multipoints$mine_fac == mine & is.na(mines_general_w_multipoints$sub_site),]$coord_is_na <- FALSE
}
 
# add mines with updated multipoints (mines_general_w_multipoints) to general_output
general_output <- bind_rows(general_output, mines_general_w_multipoints)

```


<br>

### 3.1 assign GADM region in which the mines with coordinates are located
```{r , echo = FALSE, message = TRUE} 


# First, assign GADM regions from the WU_GADM_log_concordance.csv file for those facilities where it is possible
# this saves a lot of processing time
general_output <- general_output %>% 
  left_join(WU_GADM_log_concordance, by = "facility_id")


# create table with most detailed GADM level available for each country
countries <- gadm_combined %>%
  select(GID_0) %>% 
  unique() %>% 
  as.vector()

max_gadm_level <- c()

for(country in countries$GID_0) {
  
  curr_max <- gadm_combined %>% 
    filter(GID_0 == country) %>% 
    select(where(~!all(is.na(.)))) %>% 
    names() %>%
    tail(1) %>% 
    str_sub(., -1) %>% 
    as.numeric()
  
  max_gadm_level <- c(max_gadm_level, curr_max)
  
}

gadm_levels <- tibble(countries$GID_0, max_gadm_level) %>% 
  rename(GID_0 = "countries$GID_0")


# get mines and add their country's max gadm level
general_gadm_added <- general_output %>% 
  filter(!(facility_id %in% WU_GADM_log_concordance$facility_id)) %>%
  mutate(GID_0 = substr(alphanumiso, 1,3)) %>% 
  left_join(., gadm_levels, by = "GID_0") %>% 
  select(facility_id, mine_fac, GID_0, max_gadm_level)

# initialize output dataframe
general_output_gadm <- tibble("facility_id" = chr(), "mine_fac" = chr(), "GID_0" = chr(), "max_gadm_level" = int(),
                              "GID_1" = chr(), "GID_2" = chr(), "GID_3" = chr(), "GID_4" = chr())

if(nrow(general_gadm_added) > 0){

  # loop over levels and join the geometries from GADM to the mines with coordinates
  # and assign GIDs corresponding to GADM polygons in which the mines are located
  # level 0 is already inluded (in GID_0)
  for(lvl in 1:4){
    
    #filter mines for the correct gadm level
    df <- general_gadm_added %>% 
      filter(max_gadm_level == lvl) 
    
    # skip to next level, if no mines with current level need to be georeferenced
    if(nrow(df) == 0) {next}
    
    # load correct level of the GADM dataset
    curr_GADM <- st_read(file, layer = paste0("level",lvl), quiet = TRUE) %>% 
      st_transform(., crs = st_crs(general_gadm_added)) %>% 
      select(., starts_with("GID"))
    
    # get unique countries
    # this makes the intersection process much faster, 
    # as the points do not have to be compared to all polygons
    countries <- df$GID_0 %>% unique
    
    for (country in countries){
      temp_df <- df %>% filter(GID_0 == country)
      temp_GADM <- curr_GADM %>% filter(GID_0 == country)
      
      country_output <- st_intersection(temp_df, temp_GADM)
      
      # export the df of this country
      if(country == countries[1]){
        level_output <- country_output
      } else {
        level_output <- bind_rows(level_output, country_output)
      }
      print(c("level:", lvl , "country:", country))
    }
    
    level_output <- level_output %>% 
      st_drop_geometry()
    
    # export the df of this level
    general_output_gadm <- bind_rows(general_output_gadm, level_output)
    
    print(c("level ", lvl, " intersected"))
  
  }
  
  
  # as the mines included mines with multipoints, one mine can intersect with multiple polygons.
  # thus they have to be summarised 
  general_output_gadm <- general_output_gadm %>% 
    select(facility_id, GID_0, GID_1, GID_2, GID_3, GID_4) %>%
    group_by(facility_id) %>% 
    summarise(
      facility_id = unique(facility_id),
      GID_0 = paste(unique(GID_0), collapse = " ; "),
      GID_1 = paste(unique(GID_1), collapse = " ; "),
      GID_2 = paste(unique(GID_2), collapse = " ; "),
      GID_3 = paste(unique(GID_3), collapse = " ; "),
      GID_4 = paste(unique(GID_4), collapse = " ; ")) %>%
    arrange(facility_id)
  
  
  
  # join the GADM regions from log file to the mines not automatically intersecting
  general_output <- general_output %>% 
    left_join(general_output_gadm, 
              by = "facility_id") %>% 
    mutate(GID_0 = ifelse(is.na(GID_0.x), GID_0.y, GID_0.x),
           GID_1 = ifelse(is.na(GID_1.x), GID_1.y, GID_1.x),
           GID_2 = ifelse(is.na(GID_2.x), GID_2.y, GID_2.x),
           GID_3 = ifelse(is.na(GID_3.x), GID_3.y, GID_3.x),
           GID_4 = ifelse(is.na(GID_4.x), GID_4.y, GID_4.x)) %>% 
    select(-ends_with(".x"), -ends_with(".y"))
  
  
  # export concordance log table that links facility ids to GADM regions.
  # this saves a lot of processing time, as the intersection of points with GADM polygons does not have to be done every time for all mines.
  WU_GADM_log_concordance <- general_output %>% 
    st_drop_geometry() %>% 
    filter(!is.na(GID_1)) %>% 
    select(facility_id, GID_0, GID_1, GID_2, GID_3, GID_4)
  
  write_csv(WU_GADM_log_concordance,
            "./01_input/02_lists_and_concordance_tables/01_detailed_data/WU_GADM_log_concordance.csv")

}

# check if any mine did not get a GADM polygon assigned
# if so, assign manually
mines_without_poly_sf <- general_output %>% 
  filter(GID_1 %>% is.na()) %>% 
  select(facility_id, mine_fac, alphanumiso, geometry, GID_1, GID_2, GID_3, GID_4) %>% 
  mutate(GID_0 = substr(alphanumiso, 1,3), .before = "GID_1") %>% 
  left_join(gadm_levels, by = "GID_0")

mines_without_poly <- mines_without_poly_sf %>%
  st_drop_geometry()

nrow(mines_without_poly)


# create/update log file containing the mines which do not intersect with GADM regions
if (file.exists("./01_input/03_log_files/01_detailed_data/log_location_not_intersecting_with_GADM.csv")){
  
  GADM_log <- read_csv("./01_input/03_log_files/01_detailed_data/log_location_not_intersecting_with_GADM.csv",
                       show_col_types = FALSE)
  
  missing_rows <- anti_join(mines_without_poly, GADM_log, 
                            by = "mine_fac")
  
  if (nrow(missing_rows) > 0) {
    
    GADM_log <- bind_rows(GADM_log, missing_rows)
    
    write_csv(GADM_log, "./01_input/03_log_files/01_detailed_data/log_location_not_intersecting_with_GADM.csv")
    
    print(c((nrow(missing_rows)), "Rows added"))
    print("You have to assign the GADM regions to the polygons manually.
          The function below helps you to plot the not intersecting polygons")
    
    } else {print("No new polygon added, file did not change")}

  } else {
  
  write_csv(mines_without_poly, "./01_input/03_log_files/01_detailed_data/log_location_not_intersecting_with_GADM.csv")
    
  print("New file created")
  print("You have to assign the GADM regions to the polygons manually.
        The function below helps you to plot the not intersecting polygons")
  
}


# visualise with the function created above (adapt rownumber)
# plot_GADM_polygons(mines_without_poly_sf[4,], 1000*100)


# join the GADM regions from log file to the mines not automatically intersecting
general_output <- general_output %>% 
  left_join(GADM_log %>% 
              select(mine_fac, GID_0, GID_1, GID_2, GID_3, GID_4), 
            by = c("mine_fac")) %>% 
  mutate(GID_0 = ifelse(is.na(GID_0.x), GID_0.y, GID_0.x),
         GID_1 = ifelse(is.na(GID_1.x), GID_1.y, GID_1.x),
         GID_2 = ifelse(is.na(GID_2.x), GID_2.y, GID_2.x),
         GID_3 = ifelse(is.na(GID_3.x), GID_3.y, GID_3.x),
         GID_4 = ifelse(is.na(GID_4.x), GID_4.y, GID_4.x)) %>% 
  select(-ends_with(".x"), -ends_with(".y"))

# # check 1
# general_output %>%
#   filter(facility_id == "COM00112.00") %>%
#   select(GID_0, GID_1, GID_2, GID_3, GID_4)
#
# # check 2
# general_output$GID_1 %>% is.na() %>% unique()

# # check 3
# general_output[general_output$GID_1 %>% is.na(),]

```

<br>

### 4. georeferencing of mines and regions that do not have coordinates in sheet_general or sheet_sub_sites

### 4.1. check if there are missing regions in the WU_GADM_concordance.xlsx file, if so update manually
```{r , echo = FALSE, message = FALSE} 

# mines without coordinates
mwoc <- mines_general_wo_coords %>% 
  select(mine_fac, country, alphanumiso, state, region, province, district, sector, location_municipality) %>% 
  st_drop_geometry() %>% 
  as.data.frame()

mwoc %>% names()

# split the regions into separate rows according to logic of the split_columns function
first_split <- split_columns(mwoc, ";") #split by ";" 
second_split <- split_columns(first_split, ",") #split by "," 


#cleaning
second_split <- second_split %>% 
  mutate(across(where(is.character), str_trim)) %>%
  mutate_all(na_if,"") %>% #replace "" with NA values
  as_tibble()
  

#create dataframe of unique regions in the detailed data set
detailed_regions <- second_split %>% 
  mutate(., GID_0 = substr(alphanumiso, 1,3), .after = location_municipality) %>% 
  select(-1) %>% 
  distinct() %>% 
  arrange(country)




# add missing regions to the WU_GADM_concordance table file, which links regions to geometries from the GADM package
if (file.exists("./01_input/02_lists_and_concordance_tables/01_detailed_data/WU_GADM_concordance.csv")){
  
  WU_GADM_concordance <- read_delim("./01_input/02_lists_and_concordance_tables/01_detailed_data/WU_GADM_concordance.csv", 
                                    delim = ";",
                                    show_col_types = FALSE ) 
  
  missing_rows <- anti_join(detailed_regions, WU_GADM_concordance, 
            by = c("country", "state", "region", "province", "district", "sector", "location_municipality"))
  
  if (nrow(missing_rows) > 0) {
    
    #add GID cols
    GID <- data.frame(matrix(ncol = 4, nrow = nrow(missing_rows)))
    colnames(GID) <- c("GID_1", "GID_2", "GID_3", "GID_4")
    
    missing_rows <- cbind(missing_rows, GID)
    WU_GADM_concordance <- rbind(WU_GADM_concordance, missing_rows) %>% 
      arrange(., country, state, district, location_municipality)
    
    write_delim(WU_GADM_concordance, "./01_input/02_lists_and_concordance_tables/01_detailed_data/WU_GADM_concordance.csv", 
                delim = ";")
    print(c((nrow(missing_rows)), "Rows added"))
    
    } else {print("No new regions added, file did not change")}

  } else {
    
  # create columns where the GID will be
  GID <- data.frame(matrix(ncol = 4, nrow = nrow(detailed_regions)))
  colnames(GID) <- c( "GID_1", "GID_2", "GID_3", "GID_4")
  
  #cbind them together
  WU_GADM_concordance <- cbind(detailed_regions, GID) %>% arrange(., country, state, district, location_municipality)

  write_delim(WU_GADM_concordance, "./01_input/02_lists_and_concordance_tables/01_detailed_data/WU_GADM_concordance.csv",
              delim = ";", show_col_types = FALSE)
  print("New file created")
  
}


```

<br>

### 4.2.1. add GADM regions to mines without coordinates
```{r , echo = FALSE, message = FALSE} 
# join the polygons to mines via the concordance table
WU_GADM_concordance <- read_delim("./01_input/02_lists_and_concordance_tables/01_detailed_data/WU_GADM_concordance.csv",
                                  delim = ";",
                                  show_col_types = FALSE)


mines_wo_georeference <- left_join(second_split, WU_GADM_concordance, 
                                   by = c("country", "alphanumiso", "state", "region", "province", 
                                          "district", "sector", "location_municipality"))


#this should return TRUE
nrow(filter(mines_wo_georeference, is.na(GID_0))) == 0

if(add_polygons == FALSE) {

  # summarize GIDs describing one mine in one row
  mines_wo_georeference_union <- mines_wo_georeference %>% 
    group_by(mine_fac) %>%
    summarize(GID_0 = paste(unique(GID_0), collapse = " ; "),
              GID_1 = paste(unique(GID_1), collapse = " ; "),
              GID_2 = paste(unique(GID_2), collapse = " ; "),
              GID_3 = paste(unique(GID_3), collapse = " ; "),
              GID_4 = paste(unique(GID_4), collapse = " ; ")) %>% 
    mutate(across(c(GID_1, GID_2, GID_3, GID_4), . %>%  str_replace(" ; NA", "")))
  

  # keep record of whether polygons were aggregated in the previous step (by st_union)
  # and if so, how many polygons there were originally
  nr_features <- mines_wo_georeference %>%
    count(mine_fac) %>% 
    rename(n_features = n)
  
  mines_wo_georeference_union <- mines_wo_georeference_union %>% left_join(., nr_features, by = "mine_fac")
  
  
  # this has to return TRUE
  nrow(mines_general_wo_coords) == nrow(mines_wo_georeference_union)
  
  
  # bind general_output and mines_with_polygons together
  mines_wo_georeference_final <- left_join(mines_general_wo_coords, 
                                           mines_wo_georeference_union, by = "mine_fac")
  
  
  # set coord_is_na to FALSE, as these mines just got georeferenced
  mines_wo_georeference_final$coord_is_na <- FALSE
  
  
  # add to general output
  general_output <- bind_rows(general_output, mines_wo_georeference_final)
  
  
  # cleaning and sorting
  general_output <- general_output %>% 
    select(-sub_site, -sub_site_other_name, -coord_is_na) %>% 
    arrange(facility_id)
  
  
  # CHECK: this has to return TRUE 
  nrow(filter(general, is.na(sub_site))) == nrow(general_output)
  nrow(general_output) + nrow(mines_sub_sites) == nrow(general)
  

}

```

<br>

### 4.2.2. OPTIONAL: add polygons to mines without coordinates
```{r , echo = FALSE, message = FALSE} 
if(add_polygons == TRUE) {

  # insert additional row when level is level 4 to circumvent lvl + 1 throwing error in the loop below
  mines_wo_georeference <- mines_wo_georeference %>% mutate(GID_5 = NA)
  
  
  # construct output df 
  mines_wo_georeference_output <- data.frame(matrix(ncol = 3, nrow = 0))
  colnames(mines_wo_georeference_output) <- c("mine_fac", "geom", "GADM_level")
  
  
  # loop over levels and join the geometries from GADM to the mines without coordinates
  for(lvl in 0:4){
    
    # load in correct level of the GADM dataset
    curr_GADM <- st_read(file, layer = paste0("level",lvl), quiet = TRUE) %>% 
      st_transform(., crs = st_crs(mines_wo_georeference))
    
    # set current GID
    curr_GID <- paste0("GID_",as.character(lvl))
    
    # set next GID
    next_GID <- paste0("GID_",as.character(lvl + 1))
    
    # filter mines where a polygon of the current level should be assigned
    curr_mines <- mines_wo_georeference[as.vector(!is.na(mines_wo_georeference[ , 10 + lvl ])), ]
    curr_mines <- curr_mines[as.vector(is.na(curr_mines[ , 11 + lvl ])), ]
    
    # join
    curr_mines <- left_join(curr_mines, curr_GADM)
  
    # cleaning
    curr_mines <- curr_mines %>% select(., mine_fac, geom, GID_0, GID_1, GID_2, GID_3, GID_4)
    
    #adding the level of the geometry
    GADM_level <- rep(lvl, nrow(curr_mines))
    curr_mines <- cbind(curr_mines, GADM_level)
  
    # append to output dataframe
    mines_wo_georeference_output <- rbind(mines_wo_georeference_output, curr_mines)
    
    print(c("level ", lvl, " joined"))
    
  }
  
  
  # aggregate polygons per mine
  mines_wo_georeference_union <- mines_wo_georeference_output %>% 
    group_by(mine_fac) %>%
    summarize(geometry = st_union(geom), 
              GADM_level = min(GADM_level),
              GID_0 = paste(unique(GID_0), collapse = " ; "),
              GID_1 = paste(unique(GID_1), collapse = " ; "),
              GID_2 = paste(unique(GID_2), collapse = " ; "),
              GID_3 = paste(unique(GID_3), collapse = " ; "),
              GID_4 = paste(unique(GID_4), collapse = " ; "))
  
  
  
  
  
  # keep record of whether polygons were aggregated in the previous step (by st_union)
  # and if so, how many polygons there were originally
  nr_features <- mines_wo_georeference %>%
    count(mine_fac) %>% 
    rename(n_features = n)
  
  mines_wo_georeference_union <- mines_wo_georeference_union %>% left_join(., nr_features, by = "mine_fac")
  
  
  # this has to return TRUE
  nrow(mines_general_wo_coords) == nrow(mines_wo_georeference_union)
  
  
  # bind general_output and mines_with_polygons together
  mines_wo_georeference_final <- left_join(st_drop_geometry(mines_general_wo_coords), 
                                           mines_wo_georeference_union, by = "mine_fac") %>% st_as_sf()
  
  
  # set coord_is_na to FALSE, as these mines just got georeferenced
  mines_wo_georeference_final$coord_is_na <- FALSE
  
  
  # add to general output
  general_output <- bind_rows(general_output, mines_wo_georeference_final)
  
  
  # cleaning and sorting
  general_output <- general_output %>% 
    select(-sub_site, -sub_site_other_name, -coord_is_na) %>% 
    arrange(facility_id)
  
  
  # CHECK: this has to return TRUE 
  nrow(filter(general, is.na(sub_site))) == nrow(general_output)
  nrow(general_output) + nrow(mines_sub_sites) == nrow(general)

}




# quick check if all polygons were assigned correctly 
if(general_output %>% 
    mutate(iso_alpha_3 = substr(alphanumiso, 1, 3)) %>% 
    filter(iso_alpha_3 != GID_0) %>% 
    nrow() > 0
) {warning("There are mines with assigned GADM regions not located in the correct country")}

```


<br>

### 5. Export data
```{r , echo = FALSE, message = FALSE} 

if (add_polygons == TRUE) {
  
  # export sheet_general as geopackage with geometries
  st_write(general_output, "./03_intermediate/01_detailed_data/sheet_general/general_with_polygons.gpkg", append = FALSE)

} else {
  
  # export sheet_general as geopackage without polygons 
  st_write(general_output, "./03_intermediate/01_detailed_data/sheet_general/07_general.gpkg", append = FALSE)
  
}

# export sheet_sub_sites as geopackage with coordinates as POINT
st_write(mines_sub_sites %>% 
           relocate(sub_site_other_name, .after = sub_site) %>%
           select(-coord_is_na) %>%
           arrange(facility_id),
         "./03_intermediate/01_detailed_data/sheet_general/sub_sites.gpkg", append = FALSE)


```
